+++ libs/sub_server/server_updates/ServerUpdateWorld.cpp
@@ -4,12 +4,12 @@
 
 #include "ServerUpdateWorld.h"
 
-uint8_t ServerUpdateWorld::getType() const {
+uint16_t ServerUpdateWorld::getType() const {
     return WORLD;
 }
 
-std::vector<uint8_t> ServerUpdateWorld::beSerialized() { //TODO: check how to serialize
-    std::vector<uint8_t> worldData;
+std::vector<uint16_t> ServerUpdateWorld::beSerialized() { //TODO: check how to serialize
+    std::vector<uint16_t> worldData;
     worldData.push_back(getType());
     worldData.insert(worldData.end(), returnData.begin(), returnData.end());
     return worldData;
