+++ libs/sub_server/server_updates/ServerUpdateWorld.h
@@ -9,12 +9,12 @@
 
 class ServerUpdateWorld : public ServerUpdate{
 public:
-    ServerUpdateWorld(uint8_t &id, uint8_t &returnCode, const uint8_t &n, std::vector<uint8_t> &returnData) :
+    ServerUpdateWorld(uint16_t &id, uint16_t &returnCode, const uint16_t &n, std::vector<uint16_t> &returnData) :
                                     ServerUpdate(id,returnCode,returnData){};
     ~ServerUpdateWorld() override = default;
 
-    uint8_t getType() const override;
-    std::vector<uint8_t> beSerialized() override;
+    uint16_t getType() const override;
+    std::vector<uint16_t> beSerialized() override;
 };
